Add files via upload
WeRateDogs : A Data Wrangling Project

by Cyril

Objectives

This is a repository for Udacity Data Analyst Project 2 (Data Wrangling). The dataset used in the project is included in this repository. The goal is to broaden my data gathering and wrangling skills as well as to produce Insights from the analysis.